rename to input-code
