+++ packages/sandbox/e2e/visual_regressions_flow.yml
@@ -88,6 +88,10 @@ appId: ${APPID}
 - runFlow:
     file: subflow/takeScreenshot.yaml
     env:
+      COMPONENT: InputCode
+- runFlow:
+    file: subflow/takeScreenshot.yaml
+    env:
       COMPONENT: ListItem
 - runFlow:
     file: subflow/takeScreenshot.yaml
