+++ packages/sandbox/src/app/sandbox/index.ts
@@ -9,7 +9,7 @@ import { CardSandbox } from './docs/card.doc';
 import { CheckboxSandbox } from './docs/checkbox.doc';
 import { DialogSandbox } from './docs/dialog.doc';
 import { IconsSandbox } from './docs/icon.doc';
-import { InputNumberSandbox } from './docs/input-number.doc';
+import { InputCodeSandbox } from './docs/input-code.doc';
 import { InputTextAreaSandbox } from './docs/input-text-area.doc';
 import { InputTextTypesSandbox } from './docs/input-text-types.doc';
 import { InputTextSandbox } from './docs/input-text.doc';
@@ -75,7 +75,7 @@ export const sandboxItems: SandBoxSectionType[] = [
       { title: 'InputText', SandBox: InputTextSandbox },
       { title: 'InputTextTypes', SandBox: InputTextTypesSandbox },
       { title: 'InputTextArea', SandBox: InputTextAreaSandbox },
-      { title: 'InputNumber', SandBox: InputNumberSandbox },
+      { title: 'InputCode', SandBox: InputCodeSandbox },
     ],
   },
   {
